feat(category-preview): useCategoryPreviewState 훅 추가 및 CategoryPreviewSection 컴포넌트 수정
